Update regression topic to include background materials from cross validation and simple linear regression project
I made this topic a hard prerequisite for cross-validation and simple linear regression so learners will go through all the materials needed for them to understand the project cross validation and simple linear regression. This topic has R-Squared resources, R-Squared is a concept that some learners have struggled with in the past.